feature: scroll edit and text widgets to bottom on refresh
This commit extends the custom attributes of text and edit widgets to
allow scrolling their content to the bottom when the widget is refreshed.
This allows "following" content updates if the widget is set to
auto-refresh - sort of like tail -f does.

Since the edit widget is scrollable by definition, just the new
scroll-to-bottom="true" custom attribute suffices to set the widget to
scroll on refresh.
The text widget takes on this new attribute, as well as the same group
of "scrolling" attributes that the vbox and hbox widgets define for
themselves.
The edit widget doesn't support pango markdown but the text widget does.

To see this new feature in action run the included example scripts.

Documentation updated. Note that I set the generic label "0.8.4+" as
the inception version of this feature.