feature: new chooser widget
The file chooser widget has been included in gtkdialog since the early
days but was shadowed by the chooser dialog, which can be invoked with
the `fileselect` function.  Consequently, Thunor didn't give the chooser
widget the kind of feature uplift that was given to other widget, and
moved the chooser widget to the miscellaneous example folder, where it
can still be found today.

However, for a specific class of applications - file viewer such as
mm_view and desktop wallpaper setters - the file chooser can be very
valuable: it is the only widget that allows embedding a full file
browser in a larger dialog.  The trouble was that the existing
implementation was not on par with the rest of gtkdialog.  This commit
fixes that and elevates the chooser widget to full citizenship. May
application developers find new ways to use it.

Now the chooser widget works consistently with the rest of gtkdialog:
- syntax[0]
- conditional actions
- standard signals
- widget signals[1]
- widget functions
- input and output directives[2]

[0] Custom attribute "when" is DEPRECATED but still recognized and
converted into attribute "signal" internally for backward compatibility.

[1] The confirm-overwrite signal isn't connected because the widget
only works in "file save" mode.
I didn't implement

[2] "auto-refresh" and "file-monitor" custom tags not implemented.
Use a hidden text widget to assist if such matters are needed.

New demo scripts and html documentation included.