new window attr: 'icon-name'
This allows the use of a gtk theme icon for the window/taskbar icon.
It works in gtk2 and gtk3. It is useful if you add a theme icon
(can and should be in the 'hicolor' theme) for your app.
This is needed for wayland support (eg sfwbar can now display
appropriate icons instead of the default) however the old
method of 'image-name' doesn't work in wayland because 'app_id'
is needed. This is why there is a call to `g_set_prgname()`